+++ AuctionTrading/AuctionTrading.Domain/Interfaces/ISeller.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using AuctionTrading.Domain.Entities;
 
-namespace AuctionTrading.Domain.Entities
+namespace AuctionTrading.Domain.Interfaces
 {
     internal interface ISeller : IVisitor
     {
